+++ src/utils/log.hpp
@@ -6,7 +6,7 @@
 /*   By: emadriga <emadriga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/11 17:24:29 by emadriga          #+#    #+#             */
-/*   Updated: 2023/05/07 23:28:49 by emadriga         ###   ########.fr       */
+/*   Updated: 2023/05/10 16:39:35 by emadriga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,4 +34,6 @@ enum e_ansi_color{
 # define LOG_ERROR(A) std::cerr << ERROR_MSG << A << std::endl
 # define LOG_COLOR(COLOR, A) std::cout << "\033[0;3" << COLOR <<"m" << A << RESET_COLOR <<std::endl
 # define LOG_BCOLOR(COLOR, A) std::cout << "\033[0;4" << COLOR <<"m" << A << RESET_COLOR <<std::endl
+# define COLOR(COLOR, A) "\033[0;3" << COLOR <<"m" << A << RESET_COLOR
+# define BCOLOR(COLOR, A) "\033[0;4" << COLOR <<"m" << A << RESET_COLOR
 #endif